feat: skip reads shorter than a window
After the windows merging, it is possible to find reads intersecting
with a region (the set of merged windows) that are shorter than the size
of the initial window. These can contribute to noise because they were
not included in the original set of reads that were involved in the
cluster analyses.