feat: improve reads assignments
Previously all the reads that led to the same set of modified indices
after the filtering phase were considered equally usable. However, the
set of modified indices (after the filtering phase) was being used to
map back to the original reads. But the original read could have more
modifications, causing the match to be impossible in these cases.

Now we track individually the filtered reads that are assigned to each
cluster, therefore it is possible to go back to the original reads and
have a reliable assignment phase and coverage values.