chore: use either tsan or asan+ubsan for tests
In order to simplify maintainability, I want to remove most of the
"parallel" layer and the `USE_TBB` flag. But TBB must be compiled with
the tsan support in order to get meaningful results, which cannot run
with other sanitizers. Therefore it is important to use the tests with
one or the others.