fix: div by 0 + conversion to long
In case a transcript is exactly one window long, the count to evaluate
how many full windows are needed is a divide by zero (because it uses
`n_windows - 1`), which is nan and it is then converted to long. But
obviously nan cannot be represented outside a floating point number.