feat: allow merging overlapping regions
Now it is possible to merge "distant" windows in a single region even if
this leads to overlapping regions. This can be useful because,
especially for combinations of large windows and small sliding window
values, the effects of multiple conformations can extend across smaller
regions with a different set of conformations (generally just one). In
this way it is easier to correlate which conformation is which, because
the merging of the windows is normally run and it is not necessary to do
it "by hand" after running Draco.