feat: support mm replicates
We can consume multiple replicates of the same analysis, potentially
improving the graph cut algorithm.

There are still many limitations, for instance the mm files must have
the same set of transcripts and they must be in the same order, but
eventually these constraints can be relaxed in the future.