feat: impl eigenvectors_to_weighted_clusters
This helper function is going to be necessary to generate a set of
weights starting from the eigenvectors of the comodification normalized
laplacian matrix. This approach relies on the fiedler vector in the case
of two clusters, otherwise it uses the kmeans to get the centroids of a
subset of eigenvectors and calculate the pairwise distances between
centroids and eigenvectors. The result, once normalized, can be used as
cluster weights.